+++ source/anymal_locomotion/anymal_locomotion/tasks/manager_based/locomotion/velocity/config/anymal_d/agents/rsl_rl_ppo_cfg.py
@@ -14,3 +14,4 @@ class AnymalDLocomotionFlatPPORunnerCfg(AnymalDFlatPPORunnerCfg):
     def __post_init__(self) -> None:
         super().__post_init__()
         self.experiment_name = "anymal_d_locomotion_v1"
+        self.max_iterations = 1000